Fix clipboard integration in zellij in vscode
Kitty required the shell command to be changed because zellij does not
expand environment variables in the copy_command, so `cb` has to be in
$PATH, but when using zellij as a shell, the file that adds
`~/.nix-profile` to the $PATH is never sourced.

No changes were required in vscode, it loads environment variables
properly, so `cb` is in $PATH for zellij.