Fix wireguard setup by switching to wg-quick
With the regular wireguard, I had to run

sudo ip link set down wg0
sudo ip link set up wg0

to get the initial handshake going. With wg-quick, that's not the
case.
I tried a few other fixes, but they didn't work.

Persistent keepalive is still a bit weird, a new handshake is only
exchanged every ~2 minutes, not 25 seconds, but whatever. It works now.

The only disadvantage is that now, the private key file is not going to
be generated automatically on the first installation (I think).
Let's see.