Lock docker images
By default, podman will just pull the images on first startup. But this
means that my configuration might break some time in the future. By
locking the images to pull during evaluation, I can guarantee that
switching back to this configuraiton will always result in the exact
same binaries being run, and even reproducing this exact state in the
future should be possible as long as dockerhub is online.