Add backups to borgbase
This requires using sops to get the key passphrase onto the target
machine. I'm not quite sure about the re-deployment strategy;
potentially I have to move the backups to a new module and comment it
out during the first deployment. It seems to me that sops would not be
able to decrypt the secrets on the target machine during the first
deployment, because I manually have to derive an age key from the newly
generated ssh key.

Maybe that just means the the first deployment only partially succeeds,
it should still build just fine.