Fix wireguard eating all outgoing traffic
I don't want to use wireguard as a full-on VPN, it should only create
tunnels to the connected peers. I didn't fully understand what the
"AllowedIPs" parameter did, so I just copied it from the wiki.

Now I learned it basically sets up a routing rule, and 0.0.0.0/0 catches
all traffic, which is what you want if you need to change your apparent
location, but not what you want if you just want to tunnel in to ssh
into your machines.