Access services directly at home
junction now acts as the main DNS server for my home, which means that
all services don't require internet access anymore if my internet is
down. It doesn't seem like roun-trip times improved, running

    time https 192.168.178.13/login Host:cloud.uhl.cx --verify=no

takes exactly the same amount of time, but for raw throughput this
should be vastly better. And it's more efficient for sure.