Allow HTTP traffic only through wireguard
Previously, it was still possible to access all services via their HTTP
ports in my home network, which made this traffic unencrypted.
This is not a problem if I trust every device on my network, which I do,
but it still didn't feel right to leave this option open.

For nextcloud, this change was trivial, but for the services managed by
podman I struggled at first, because the firewall rules actually don't
do anything. Podman makes its own firewall rules based on the `ports`
attribute, so I had to restrict by incoming IP there, not interface.

This change still allows devices on the wireguard network to access the
sites unencrypted, but I trust all of those 100%.