Create mkDeploy
This helps make deployments easier with flakes. In the context of this
flake, you can run builds similar to how flakey-profile does it:

nix run path:.#deploy.junction.build

There's also a short help available via:

nix run path:.#deploy

Deployments are remote if the target's hostname is not the same as the
native hostname, otherwise they work just like nixos-rebuild.

I'll use this for a while and see how it feels.