Upgrade to Nextcloud 28
This upgrade was blocked by News still only having an unstable build
available for Nextcloud versions higher than 27:

- https://apps.nextcloud.com/apps/news
- https://github.com/nextcloud/news/issues/2503

After reading a bit, I decided the rewrite was stable enough for me.
The upgrade initially disabled the app, but running these commands gave
me the unstable build and enabled it just fine:

    nextcloud-occ app:update --allow-unstable -- news
    nextcloud-occ app:enable -- news

I also had to remove the Right-Click App (which has been built in since
a long time, it seems), and manually launch the Recognize App's repair
mechanism after the first attempt failed.

    nextcloud-occ maintenance:repair
    nextcloud-occ recognize:download-models

The additional configuration option regarding the maintenance window was
a reaction to a warning in the admin settings.

Apart from that, everything is working great now!