Allow gateway to build remotely on junction
To even create this change, I first had to deploy to gateway, then get
the public key, insert it into the config, and then deploy to junction.

After that, I did one login from gateway to junction to make sure the
known-hosts file was up-to-date:

    felix@horse$ ssh junction ssh-keygen -lf /etc/ssh/ssh_host_ed25519_key.pub
    256 SHA256:c99fonm0ruZ/eNoU/vgHba3fZibL2lL43DD3wlZf7gU root@junction (ED25519)
    felix@horse$ ssh gateway
    felix@gateway$ sudo -E zsh
    root@gateway$ ssh remote-build@junction
    The authenticity of host 'junction (10.100.0.13)' can't be established.
    ED25519 key fingerprint is SHA256:c99fonm0ruZ/eNoU/vgHba3fZibL2lL43DD3wlZf7gU.
    This key is not known by any other names.
    Are you sure you want to continue connecting (yes/no/[fingerprint])? SHA256:c99fonm0ruZ/eNoU/vgHba3fZibL2lL43DD3wlZf7gU
    Warning: Permanently added 'junction' (ED25519) to the list of known hosts.
    [remote-build@junction:~]$

And that's it. This should help me avoid compilation of Lix on gateway
after deploying it to junction first.